requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
docutils==0.19
idna==3.4
Kivy==2.1.0
Kivy-Garden==0.1.5
Pygments==2.13.0
PyYAML==6.0
requests==2.28.1
urllib3==1.26.12
